% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunexomePeak.R
\name{RunexomePeak}
\alias{RunexomePeak}
\title{RunexomePeak: Process Exome Peak Data}
\usage{
RunexomePeak(
  InputBAM =
    "~/m6A_calling_strategy/6_alignment_star/R2_bam/mESC_NEB_WT_rep2_Input_STAR.Aligned.out.SortedByCoord.unique.deduplicate.R2.bam",
  RIPBAM =
    "~/m6A_calling_strategy/6_alignment_star/R2_bam/mESC_NEB_WT_rep2_RIP_STAR.Aligned.out.SortedByCoord.unique.deduplicate.R2.bam",
  outdir = "~/m6A_calling_strategy/exomePeak",
  prefix = "mESC_NEB_WT_rep2_UD_test",
  gtf = "~/genome_db/mm39/gencode.GRCm39.vM30.annotation.gtf",
  window.width = 50,
  sliding.step = 50,
  peak_cutoff_pvalue = c(1, 0.05, 0.01, 1e-05, 1e-06),
  peak_cutoff_fdr = c(NA, 1, 0.05, 0.01, 0.001),
  fold_enrichment = 1
)
}
\arguments{
\item{...}{Arguments.}
}
\value{
Processed data.
}
\description{
Brief description.
}
